Implement dependency manager for MySQL and SQLite
Some platforms such as Nukkit do not have some of these jars shaded into their classpath. Due to this, we have to install external libraries. What this does is downloads the missing jar from the internet, adds it to a libraries folder, then adds the downloaded jar(s) to the classpath.

Sponge is intentionally exempt from this because it has its own MySQL and SQLite drivers. This is also evident in SQLSerializer as well to prevent a ClassNotFoundException.